+++ src/print/mod.rs
@@ -22,7 +22,7 @@ use itertools::Itertools as _;
 use crate::func_at::FuncAt;
 use crate::print::multiversion::Versions;
 use crate::qptr::{self, QPtrAttr, QPtrMemUsage, QPtrMemUsageKind, QPtrOp, QPtrUsage};
-use crate::visit::{DynVisit, InnerVisit, Visit, Visitor};
+use crate::visit::{InnerVisit, Visit, Visitor};
 use crate::{
     cfg, spv, AddrSpace, Attr, AttrSet, AttrSetDef, Const, ConstCtor, ConstDef, Context,
     ControlNode, ControlNodeDef, ControlNodeKind, ControlNodeOutputDecl, ControlRegion,
@@ -59,22 +59,17 @@ pub struct Plan<'a> {
 
     /// When visiting module-stored nodes, the [`Module`] is needed to map the
     /// [`Node`] to the (per-version) definition, which is then stored in the
-    /// (per-version) [`FxHashMap`] within `per_version_name_and_node_defs`.
+    /// (per-version) `node_defs` map.
     current_module: Option<&'a Module>,
 
     /// Versions allow comparing multiple copies of the same e.g. [`Module`],
     /// with definitions sharing a [`Node`] key being shown together.
     ///
-    /// Each `per_version_name_and_node_defs` entry contains a "version" with:
-    /// * a descriptive name (e.g. the name of a pass that produced that version)
-    ///   * the name is left empty in the default single-version mode
-    /// * its [`Node`] definitions (dynamic via the [`DynNodeDef`] helper trait)
-    ///
     /// Specific [`Node`]s may be present in only a subset of versions, and such
     /// a distinction will be reflected in the output.
     ///
-    /// For [`Node`] collection, the last entry consistutes the "active" version.
-    per_version_name_and_node_defs: Vec<(String, FxHashMap<Node, &'a dyn DynNodeDef<'a>>)>,
+    /// For [`Node`] collection, `versions.last()` constitutes the "active" one.
+    versions: Vec<PlanVersion<'a>>,
 
     /// Merged per-[`Use`] counts across all versions.
     ///
@@ -84,20 +79,26 @@ pub struct Plan<'a> {
     use_counts: FxIndexMap<Use, usize>,
 }
 
-/// Helper for printing a mismatch error between two nodes (e.g. types), while
-/// taking advantage of the print infrastructure that will print all dependencies.
-pub struct ExpectedVsFound<E, F> {
-    pub expected: E,
-    pub found: F,
+/// One version of a multi-version [`Plan`] (see also its `versions` field),
+/// or the sole one (in the single-version mode).
+struct PlanVersion<'a> {
+    /// Descriptive name for this version (e.g. the name of a pass that produced it),
+    /// or left empty (in the single-version mode).
+    name: String,
+
+    /// Definitions for all the [`Node`]s which may need to be printed later
+    /// (with the exception of [`Node::AllCxInterned`], which is special-cased).
+    node_defs: FxHashMap<Node, NodeDef<'a>>,
+
+    /// Either a whole [`Module`], or some other printable type passed to
+    /// [`Plan::for_root`]/[`Plan::for_versions`], which gets printed last,
+    /// after all of its dependencies (making up the rest of the [`Plan`]).
+    root: &'a dyn Print<Output = pretty::Fragment>,
 }
 
 /// Print [`Plan`] top-level entry, an effective reification of SPIR-T's implicit DAG.
 #[derive(Copy, Clone, PartialEq, Eq, Hash)]
 enum Node {
-    /// Either a whole [`Module`], or some other printable type passed to
-    /// [`Plan::for_root`] (e.g. [`ExpectedVsFound`]).
-    Root,
-
     /// Definitions for all [`CxInterned`] that need them, grouped together.
     AllCxInterned,
 
@@ -113,8 +114,6 @@ enum Node {
 impl Node {
     fn category(self) -> Result<&'static str, &'static str> {
         match self {
-            Self::Root => Err("Node::Root"),
-
             Self::AllCxInterned => Err("Node::AllCxInterned"),
 
             // FIXME(eddyb) these don't have the same kind of `{category}{idx}`
@@ -128,8 +127,17 @@ impl Node {
     }
 }
 
-/// Helper for [`Node::AllCxInterned`]'s definition, to  be used in `node_defs`.
-struct AllCxInterned;
+/// Definition of a [`Node`] (i.e. a reference pointing into a [`Module`]).
+///
+/// Note: [`Node::AllCxInterned`] does *not* have its own `NodeDef` variant,
+/// as it *must* be specially handled instead.
+#[derive(Copy, Clone, derive_more::From)]
+enum NodeDef<'a> {
+    ModuleDialect(&'a ModuleDialect),
+    ModuleDebugInfo(&'a ModuleDebugInfo),
+    GlobalVar(&'a GlobalVarDecl),
+    Func(&'a FuncDecl),
+}
 
 /// Anything interned in [`Context`], that might need to be printed once
 /// (as part of [`Node::AllCxInterned`]) and referenced multiple times.
@@ -147,23 +155,6 @@ impl CxInterned {
         }
     }
 }
-
-/// A [`Print`] `Output` type that splits the attributes from the main body of the
-/// definition, allowing additional processing before they get concatenated.
-#[derive(Default)]
-pub struct AttrsAndDef {
-    pub attrs: pretty::Fragment,
-
-    /// Definition that typically looks like one of these cases:
-    /// * ` = ...` for `name = ...`
-    /// * `(...) {...}` for `name(...) {...}` (i.e. functions)
-    ///
-    /// Where `name` is added later (i.e. between `attrs` and `def_without_name`).
-    pub def_without_name: pretty::Fragment,
-}
-
-trait DynNodeDef<'a>: DynVisit<'a, Plan<'a>> + Print<Output = AttrsAndDef> {}
-impl<'a, T: DynVisit<'a, Plan<'a>> + Print<Output = AttrsAndDef>> DynNodeDef<'a> for T {}
 
 #[derive(Copy, Clone, PartialEq, Eq, Hash)]
 enum Use {
@@ -240,16 +231,9 @@ impl<'a> Plan<'a> {
     // FIXME(eddyb) consider renaming this and removing the `for_module` shorthand.
     pub fn for_root(
         cx: &'a Context,
-        root: &'a (impl DynVisit<'a, Plan<'a>> + Print<Output = AttrsAndDef>),
+        root: &'a (impl Visit + Print<Output = pretty::Fragment>),
     ) -> Self {
-        let mut plan = Self {
-            cx,
-            current_module: None,
-            per_version_name_and_node_defs: vec![(String::new(), FxHashMap::default())],
-            use_counts: FxIndexMap::default(),
-        };
-        plan.use_node(Node::Root, root);
-        plan
+        Self::for_versions(cx, [("", root)])
     }
 
     /// Create a [`Plan`] with all of `module`'s contents.
@@ -261,7 +245,7 @@ impl<'a> Plan<'a> {
 
     /// Create a [`Plan`] that combines [`Plan::for_root`] from each version.
     ///
-    /// Each version has a string, which should contain a descriptive name
+    /// Each version also has a string, which should contain a descriptive name
     /// (e.g. the name of a pass that produced that version).
     ///
     /// While the roots (and their dependencies) can be entirely unrelated, the
@@ -273,22 +257,25 @@ impl<'a> Plan<'a> {
         versions: impl IntoIterator<
             Item = (
                 impl Into<String>,
-                &'a (impl DynVisit<'a, Plan<'a>> + Print<Output = AttrsAndDef> + 'a),
+                &'a (impl Visit + Print<Output = pretty::Fragment> + 'a),
             ),
         >,
     ) -> Self {
         let mut plan = Self {
             cx,
             current_module: None,
-            per_version_name_and_node_defs: vec![],
+            versions: vec![],
             use_counts: FxIndexMap::default(),
         };
         for (version_name, version_root) in versions {
             let mut combined_use_counts = mem::take(&mut plan.use_counts);
-            plan.per_version_name_and_node_defs
-                .push((version_name.into(), FxHashMap::default()));
+            plan.versions.push(PlanVersion {
+                name: version_name.into(),
+                node_defs: FxHashMap::default(),
+                root: version_root,
+            });
 
-            plan.use_node(Node::Root, version_root);
+            version_root.visit_with(&mut plan);
 
             // Merge use counts (from second version onward).
             if !combined_use_counts.is_empty() {
@@ -298,15 +285,6 @@ impl<'a> Plan<'a> {
                 }
                 plan.use_counts = combined_use_counts;
             }
-        }
-
-        // HACK(eddyb) avoid having `Node::Root` before any of its dependencies,
-        // but without having to go through the non-trivial approach of properly
-        // interleaving all the nodes, from each version, together.
-        {
-            let (map, k) = (&mut plan.use_counts, Use::Node(Node::Root));
-            // FIXME(eddyb) this is effectively a "rotate" operation.
-            map.shift_remove(&k).map(|v| map.insert(k, v));
         }
 
         plan
@@ -333,7 +311,10 @@ impl<'a> Plan<'a> {
         }
 
         // Group all `CxInterned`s in a single top-level `Node`.
-        self.use_node(Node::AllCxInterned, &AllCxInterned);
+        *self
+            .use_counts
+            .entry(Use::Node(Node::AllCxInterned))
+            .or_default() += 1;
 
         *self.use_counts.entry(use_kind).or_default() += 1;
     }
@@ -342,31 +323,49 @@ impl<'a> Plan<'a> {
     ///
     /// Only the first call recurses into the definition, subsequent calls only
     /// update its (internally tracked) "use count".
-    fn use_node(&mut self, node: Node, node_def: &'a dyn DynNodeDef<'a>) {
+    fn use_node<D: Visit>(&mut self, node: Node, node_def: &'a D)
+    where
+        NodeDef<'a>: From<&'a D>,
+    {
         if let Some(use_count) = self.use_counts.get_mut(&Use::Node(node)) {
             *use_count += 1;
             return;
         }
 
-        let (_, node_defs) = self.per_version_name_and_node_defs.last_mut().unwrap();
-        match node_defs.entry(node) {
+        let current_version = self.versions.last_mut().unwrap();
+        match current_version.node_defs.entry(node) {
             Entry::Occupied(entry) => {
-                let dyn_data_ptr = |r| (r as *const dyn DynNodeDef<'_>).cast::<()>();
+                let old_ptr_eq_new = match (*entry.get(), NodeDef::from(node_def)) {
+                    (NodeDef::ModuleDialect(old), NodeDef::ModuleDialect(new)) => {
+                        std::ptr::eq(old, new)
+                    }
+                    (NodeDef::ModuleDebugInfo(old), NodeDef::ModuleDebugInfo(new)) => {
+                        std::ptr::eq(old, new)
+                    }
+                    (NodeDef::GlobalVar(old), NodeDef::GlobalVar(new)) => std::ptr::eq(old, new),
+                    (NodeDef::Func(old), NodeDef::Func(new)) => std::ptr::eq(old, new),
+                    _ => false,
+                };
+
+                // HACK(eddyb) this avoids infinite recursion - we can't insert
+                // into `use_counts` before `node_def.visit_with(self)` because
+                // we want dependencies to come before dependends, so recursion
+                // from the visitor (recursive `Func`s, or `visit_foo` calling
+                // `use_node` which calls the same `visit_foo` method again)
+                // ends up here, and we have to both allow it and early-return.
                 assert!(
-                    std::ptr::eq(dyn_data_ptr(*entry.get()), dyn_data_ptr(node_def)),
+                    old_ptr_eq_new,
                     "print: same `{}` node has multiple distinct definitions in `Plan`",
                     node.category().unwrap_or_else(|s| s)
                 );
-
-                // Avoid infinite recursion for e.g. recursive functions.
                 return;
             }
             Entry::Vacant(entry) => {
-                entry.insert(node_def);
+                entry.insert(NodeDef::from(node_def));
             }
         }
 
-        node_def.dyn_visit_with(self);
+        node_def.visit_with(self);
 
         *self.use_counts.entry(Use::Node(node)).or_default() += 1;
     }
@@ -520,25 +519,6 @@ impl<'a> Visitor<'a> for Plan<'a> {
     }
 }
 
-impl<E: Visit, F: Visit> Visit for ExpectedVsFound<E, F> {
-    fn visit_with<'a>(&'a self, visitor: &mut impl Visitor<'a>) {
-        self.inner_visit_with(visitor);
-    }
-}
-
-impl<E: Visit, F: Visit> InnerVisit for ExpectedVsFound<E, F> {
-    fn inner_visit_with<'a>(&'a self, visitor: &mut impl Visitor<'a>) {
-        let Self { expected, found } = self;
-
-        expected.visit_with(visitor);
-        found.visit_with(visitor);
-    }
-}
-
-impl Visit for AllCxInterned {
-    fn visit_with<'a>(&'a self, _visitor: &mut impl Visitor<'a>) {}
-}
-
 // FIXME(eddyb) make max line width configurable.
 const MAX_LINE_WIDTH: usize = 120;
 
@@ -565,11 +545,11 @@ impl Plan<'_> {
     ) {
         let printer = Printer::new(self);
         (
-            self.print_with_node_filter(&printer, |node| !matches!(node, Node::Root))
+            self.print_all_nodes_and_or_root(&printer, true, false)
                 .map_pretty_fragments(|fragment| {
                     fragment.layout_with_max_line_width(MAX_LINE_WIDTH)
                 }),
-            self.print_with_node_filter(&printer, |node| matches!(node, Node::Root))
+            self.print_all_nodes_and_or_root(&printer, false, true)
                 .map_pretty_fragments(|fragment| {
                     fragment.layout_with_max_line_width(MAX_LINE_WIDTH)
                 }),
@@ -628,7 +608,7 @@ impl<'a> Printer<'a> {
 
                 // HACK(eddyb) these are "global" to the whole print `Plan`.
                 if let Use::Node(
-                    Node::Root | Node::AllCxInterned | Node::ModuleDialect | Node::ModuleDebugInfo,
+                    Node::AllCxInterned | Node::ModuleDialect | Node::ModuleDebugInfo,
                 ) = use_kind
                 {
                     return (
@@ -706,10 +686,7 @@ impl<'a> Printer<'a> {
                         Use::Node(Node::Func(_)) => &mut ac.funcs,
 
                         Use::Node(
-                            Node::Root
-                            | Node::AllCxInterned
-                            | Node::ModuleDialect
-                            | Node::ModuleDebugInfo,
+                            Node::AllCxInterned | Node::ModuleDialect | Node::ModuleDebugInfo,
                         )
                         | Use::ControlRegionLabel(_)
                         | Use::ControlRegionInput { .. }
@@ -782,19 +759,16 @@ impl<'a> Printer<'a> {
                 }
             };
 
-            let func_def_bodies_across_versions = plan
-                .per_version_name_and_node_defs
-                .iter()
-                .filter_map(|(_, node_defs)| {
-                    match node_defs.get(&Node::Func(func))?.downcast_as_func_decl()? {
-                        FuncDecl {
-                            def: DeclDef::Present(func_def_body),
-                            ..
-                        } => Some(func_def_body),
+            let func_def_bodies_across_versions = plan.versions.iter().filter_map(|version| {
+                match version.node_defs.get(&Node::Func(func))? {
+                    NodeDef::Func(FuncDecl {
+                        def: DeclDef::Present(func_def_body),
+                        ..
+                    }) => Some(func_def_body),
 
-                        _ => None,
-                    }
-                });
+                    _ => None,
+                }
+            });
 
             for func_def_body in func_def_bodies_across_versions {
                 let visit_region = |func_at_region: FuncAt<'_, ControlRegion>| {
@@ -1102,6 +1076,20 @@ impl<'a> Printer<'a> {
     }
 }
 
+/// A [`Print`] `Output` type that splits the attributes from the main body of the
+/// definition, allowing additional processing before they get concatenated.
+#[derive(Default)]
+pub struct AttrsAndDef {
+    pub attrs: pretty::Fragment,
+
+    /// Definition that typically looks like one of these cases:
+    /// * ` = ...` for `name = ...`
+    /// * `(...) {...}` for `name(...) {...}` (i.e. functions)
+    ///
+    /// Where `name` is added later (i.e. between `attrs` and `def_without_name`).
+    pub def_without_name: pretty::Fragment,
+}
+
 impl AttrsAndDef {
     /// Concat `attrs`, `name` and `def_without_name` into a [`pretty::Fragment`],
     /// effectively "filling in" the `name` missing from `def_without_name`.
@@ -1183,35 +1171,11 @@ impl AttrsAndDef {
 }
 
 pub trait Print {
+    // FIXME(eddyb) maybe remove `type Output;` flexibility by having two traits
+    // instead of one? (a method that returns `self.attrs` would allow for some
+    // automation, and remove a lot of the noise that `AttrsAndDef` adds).
     type Output;
     fn print(&self, printer: &Printer<'_>) -> Self::Output;
-
-    // HACK(eddyb) this is only ever implemented by `FuncDecl`, to allow for
-    // `Printer::new` to compute its per-function indices. A better replacement
-    // could eventually be `fn setup_printer(&self, printer: &mut Printer)`.
-    fn downcast_as_func_decl(&self) -> Option<&FuncDecl> {
-        None
-    }
-}
-
-impl<E: Print<Output = pretty::Fragment>, F: Print<Output = pretty::Fragment>> Print
-    for ExpectedVsFound<E, F>
-{
-    type Output = AttrsAndDef;
-    fn print(&self, printer: &Printer<'_>) -> AttrsAndDef {
-        let Self { expected, found } = self;
-
-        AttrsAndDef {
-            attrs: pretty::Fragment::default(),
-            def_without_name: pretty::Fragment::new([
-                "expected: ".into(),
-                expected.print(printer),
-                pretty::Node::ForceLineSeparation.into(),
-                "found: ".into(),
-                found.print(printer),
-            ]),
-        }
-    }
 }
 
 impl Use {
@@ -1328,63 +1292,74 @@ impl Print for Func {
 impl Print for Plan<'_> {
     type Output = Versions<pretty::Fragment>;
     fn print(&self, printer: &Printer<'_>) -> Versions<pretty::Fragment> {
-        self.print_with_node_filter(printer, |_| true)
+        self.print_all_nodes_and_or_root(printer, true, true)
     }
 }
 
 impl Plan<'_> {
-    fn print_with_node_filter(
+    fn print_all_nodes_and_or_root(
         &self,
         printer: &Printer<'_>,
-        filter: impl Fn(&Node) -> bool,
+        print_all_nodes: bool,
+        print_root: bool,
     ) -> Versions<pretty::Fragment> {
-        let num_versions = self.per_version_name_and_node_defs.len();
-        let per_node_versions_with_repeat_count = printer
+        enum NodeOrRoot {
+            Node(Node),
+            Root,
+        }
+
+        let all_nodes = printer
             .use_styles
             .keys()
             .filter_map(|&use_kind| match use_kind {
                 Use::Node(node) => Some(node),
                 _ => None,
             })
-            .filter(filter)
-            .map(|node| -> SmallVec<[_; 1]> {
-                if num_versions == 0 {
-                    return [].into_iter().collect();
+            .map(NodeOrRoot::Node);
+        let root = [NodeOrRoot::Root].into_iter();
+        let all_nodes_and_or_root = Some(all_nodes)
+            .filter(|_| print_all_nodes)
+            .into_iter()
+            .flatten()
+            .chain(Some(root).filter(|_| print_root).into_iter().flatten());
+
+        let per_node_versions_with_repeat_count =
+            all_nodes_and_or_root.map(|node_or_root| -> SmallVec<[_; 1]> {
+                // Only print `Node::AllCxInterned` once (it doesn't really have
+                // per-version node definitions in the first place, anyway).
+                if let NodeOrRoot::Node(node @ Node::AllCxInterned) = node_or_root {
+                    node.category().unwrap_err();
+
+                    return [(CxInterned::print_all(printer), self.versions.len())]
+                        .into_iter()
+                        .collect();
                 }
 
-                let name = if node.category().is_err() {
-                    pretty::Fragment::default()
-                } else {
-                    Use::Node(node).print_as_def(printer)
-                };
-
-                // Avoid printing `AllCxInterned` more than once, it doesn't
-                // really have per-version node definitions in the first place.
-                if let Node::AllCxInterned = node {
-                    // FIXME(eddyb) maybe make `DynNodeDef` `Any`-like, to be
-                    // able to assert that all per-version defs are identical.
-                    return [(
-                        AllCxInterned.print(printer).insert_name_before_def(name),
-                        num_versions,
-                    )]
-                    .into_iter()
-                    .collect();
-                }
-
-                self.per_version_name_and_node_defs
+                self.versions
                     .iter()
-                    .map(move |(_, node_defs)| {
-                        node_defs
-                            .get(&node)
-                            .map(|def| def.print(printer).insert_name_before_def(name.clone()))
-                            .unwrap_or_default()
+                    .map(move |version| match node_or_root {
+                        NodeOrRoot::Node(node) => {
+                            // HACK(eddyb) the special case was handled above.
+                            node.category().unwrap();
+
+                            let name = if node.category().is_err() {
+                                pretty::Fragment::default()
+                            } else {
+                                Use::Node(node).print_as_def(printer)
+                            };
+                            version
+                                .node_defs
+                                .get(&node)
+                                .map(|def| def.print(printer).insert_name_before_def(name.clone()))
+                                .unwrap_or_default()
+                        }
+                        NodeOrRoot::Root => version.root.print(printer),
                     })
                     .dedup_with_count()
                     .map(|(repeat_count, fragment)| {
                         // FIXME(eddyb) consider rewriting intra-func anchors
-                        // here, post-deduplication, to be unique per-version.
-                        // Additionally, a diff algorithm could be employed, to
-                        // annotate the changes between versions.
+                        // here, post-deduplication, to be unique per-version,
+                        // though `multiversion` should probably handle it.
 
                         (fragment, repeat_count)
                     })
@@ -1392,7 +1367,7 @@ impl Plan<'_> {
             });
 
         // Unversioned, flatten the nodes.
-        if num_versions == 1 && self.per_version_name_and_node_defs[0].0.is_empty() {
+        if self.versions.len() == 1 && self.versions[0].name.is_empty() {
             Versions::Single(pretty::Fragment::new(
                 per_node_versions_with_repeat_count
                     .map(|mut versions_with_repeat_count| {
@@ -1408,11 +1383,7 @@ impl Plan<'_> {
             ))
         } else {
             Versions::Multiple {
-                version_names: self
-                    .per_version_name_and_node_defs
-                    .iter()
-                    .map(|(name, _)| name.clone())
-                    .collect(),
+                version_names: self.versions.iter().map(|v| v.name.clone()).collect(),
                 per_row_versions_with_repeat_count: per_node_versions_with_repeat_count.collect(),
             }
         }
@@ -1420,13 +1391,13 @@ impl Plan<'_> {
 }
 
 impl Print for Module {
-    type Output = AttrsAndDef;
-    fn print(&self, printer: &Printer<'_>) -> AttrsAndDef {
+    type Output = pretty::Fragment;
+    fn print(&self, printer: &Printer<'_>) -> pretty::Fragment {
         if self.exports.is_empty() {
-            return AttrsAndDef::default();
+            return pretty::Fragment::default();
         }
 
-        let exports = pretty::Fragment::new([
+        pretty::Fragment::new([
             printer.declarative_keyword_style().apply("export").into(),
             " ".into(),
             pretty::join_comma_sep(
@@ -1445,11 +1416,18 @@ impl Print for Module {
                     }),
                 "}",
             ),
-        ]);
+        ])
+    }
+}
 
-        AttrsAndDef {
-            attrs: pretty::Fragment::default(),
-            def_without_name: exports,
+impl Print for NodeDef<'_> {
+    type Output = AttrsAndDef;
+    fn print(&self, printer: &Printer<'_>) -> AttrsAndDef {
+        match self {
+            Self::ModuleDialect(dialect) => dialect.print(printer),
+            Self::ModuleDebugInfo(debug_info) => debug_info.print(printer),
+            Self::GlobalVar(gv_decl) => gv_decl.print(printer),
+            Self::Func(func_decl) => func_decl.print(printer),
         }
     }
 }
@@ -1660,9 +1638,8 @@ impl Print for Exportee {
     }
 }
 
-impl Print for AllCxInterned {
-    type Output = AttrsAndDef;
-    fn print(&self, printer: &Printer<'_>) -> AttrsAndDef {
+impl CxInterned {
+    fn print_all(printer: &Printer<'_>) -> pretty::Fragment {
         let fragments = printer
             .use_styles
             .iter()
@@ -1697,10 +1674,7 @@ impl Print for AllCxInterned {
                 "\n\n".into()
             });
 
-        AttrsAndDef {
-            attrs: pretty::Fragment::default(),
-            def_without_name: pretty::Fragment::new(fragments),
-        }
+        pretty::Fragment::new(fragments)
     }
 }
 
@@ -1775,9 +1749,7 @@ impl Print for Attr {
                                 }
                             };
 
-                            let mut printed_message = message
-                                .print(printer)
-                                .insert_name_before_def(pretty::Fragment::default());
+                            let mut printed_message = message.print(printer);
 
                             // HACK(eddyb) apply the right style to all the plain
                             // text parts of the already-printed message.
@@ -1916,28 +1888,20 @@ impl Print for Attr {
 }
 
 impl Print for Vec<DiagMsgPart> {
-    // HACK(eddyb) this is only `AttrsAndDef` so it can be pretty-printed as a
-    // root - but ideally `DynNodeDef` should allow `pretty::Fragment` too.
-    type Output = AttrsAndDef;
-
-    fn print(&self, printer: &Printer<'_>) -> AttrsAndDef {
-        let msg = pretty::Fragment::new(self.iter().map(|part| match part {
+    type Output = pretty::Fragment;
+    fn print(&self, printer: &Printer<'_>) -> pretty::Fragment {
+        pretty::Fragment::new(self.iter().map(|part| match part {
             DiagMsgPart::Plain(text) => pretty::Node::Text(text.clone()).into(),
             DiagMsgPart::Attrs(attrs) => attrs.print(printer),
             DiagMsgPart::Type(ty) => ty.print(printer),
             DiagMsgPart::Const(ct) => ct.print(printer),
             DiagMsgPart::QPtrUsage(usage) => usage.print(printer),
-        }));
-        AttrsAndDef {
-            attrs: pretty::Fragment::default(),
-            def_without_name: msg,
-        }
+        }))
     }
 }
 
 impl Print for QPtrUsage {
     type Output = pretty::Fragment;
-
     fn print(&self, printer: &Printer<'_>) -> pretty::Fragment {
         match self {
             QPtrUsage::Handles(qptr::shapes::Handle::Opaque(ty)) => ty.print(printer),
@@ -1957,7 +1921,6 @@ impl Print for QPtrUsage {
 
 impl Print for QPtrMemUsage {
     type Output = pretty::Fragment;
-
     fn print(&self, printer: &Printer<'_>) -> pretty::Fragment {
         // FIXME(eddyb) should this be a helper on `Printer`?
         let num_lit = |x: u32| printer.numeric_literal_style().apply(format!("{x}")).into();
@@ -2548,10 +2511,6 @@ impl Print for FuncDecl {
             attrs: attrs.print(printer),
             def_without_name,
         }
-    }
-
-    fn downcast_as_func_decl(&self) -> Option<&FuncDecl> {
-        Some(self)
     }
 }